get products from g2a

+++ Ecommerce.UI/Shared/Product/PriceLimit.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 
 namespace Ecommerce.UI.Shared.Product
@@ -11,7 +12,7 @@ namespace Ecommerce.UI.Shared.Product
         public int Id { get; set; }
         public string IdProduto { get; set; }
         public Product Produto { get; set; }
-        public int? max { get; set; } // Maximum price (null if not defined)
-        public int? mim { get; set; } // Minimum price (null if not defined)
+        public decimal? max { get; set; } // Maximum price (null if not defined)
+        public decimal? mim { get; set; } // Minimum price (null if not defined)
     }
 }
